+++ src/libraries/resource_network/NodeProperties.h
@@ -70,7 +70,7 @@ public:
     void Update();
     void Finalize();
     void Push( int quantity, unsigned long resourceId );
-    void SetModifier( unsigned int modifier );
+    void SetModifier( float modifier );
     void AddConsumption( unsigned long resourceId, unsigned int consumption );
     bool NeedUpdate() const;
     float GetFunctionalState() const;
