+++ src/libraries/simulation_kernel/Entities/Inhabitants/MIL_Inhabitant.h
@@ -75,7 +75,7 @@ public:
 
     //! @name From MIL_StructuralStateNotifier_ABC
     //@{
-    virtual void NotifyStructuralStateChanged( unsigned int structuralState, const MIL_Object_ABC& object ) {}
+    virtual void NotifyStructuralStateChanged( float structuralState, const MIL_Object_ABC& object ) {}
     virtual void NotifyFired();
     //@}
