+++ src/libraries/simulation_kernel/Entities/Objects/MIL_StructuralStateNotifier_ABC.h
@@ -31,7 +31,7 @@ public:
 
     //! @name Operations
     //@{
-    virtual void NotifyStructuralStateChanged( unsigned int structuralState, const MIL_Object_ABC& object ) = 0;
+    virtual void NotifyStructuralStateChanged( float structuralState, const MIL_Object_ABC& object ) = 0;
     virtual void NotifyFired() = 0;
     //@}
 };
